refactor: introduce `test/helpers` directory
In order to distinguish between test files and test helper files.
