+++ test/helpers/index.js
@@ -1,0 +1,7 @@
+import assertThrows from './assert-throws'
+import exec from './exec'
+
+export {
+  assertThrows,
+  exec,
+}
